requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
minio
pyspark
apache-airflow-providers-apache-spark
apache-airflow-providers-postgres
pandas
psycopg2-binary
